+++ index.js
@@ -80,6 +80,8 @@ Hawkejs.setMethod(function afterInit() {
 
 	// Require these files in the browser only
 	this.load('lib/client/isvisible.js', {server: false});
+	this.load('lib/client/register_element.js', {server: false});
+	this.load('lib/class/custom_element.js', {server: false});
 	this.load('lib/client/history.js', {server: false});
 	this.load('lib/class/hawkejs-client.js', {server: false});
 	this.load('lib/class/scene.js', {server: false});